JT-3784 Fix some bugs related to test score display

+++ lib/school_profiles/data_point.rb
@@ -59,7 +59,7 @@ module SchoolProfiles
     end
 
     def float_value
-      value.to_s.scan(/[0-9.]+/).first.to_i
+      value.to_s.scan(/[0-9.]+/).first.to_f
     end
 
     def format
